Create game module and domain entities
Create game module that will be used to process all information related to the fps game.

Also create all the domain entities that will store information parsed from the logs and the rankings.

Game - Contains all information related to the matches

Player - Information related to a player in a match

Frag - All frags done by a player

Ranking - Stores ranking information related to a game. If its a global ranking, game id will be empty.

PlayerRanking - Position and awards of a player in the ranking.