Add team support
Add team support to the log parser. A team is idenfied by a single letter after the player's name, between parenthesis.

Example:

`23/04/2019 15:36:04 - Roman(A) killed Nick(B) using M16`

The logs can be read with or without teams.

The team information is used to check if a kill is friendly fire or not.