Add killstreak to rankings
Calculate killstreak in ranking service.

A killstreak is the number of frags a player has between deaths. The calculation is made using frags and deaths timestamp.
While frags time is less than death time, keep increasing killstreak. When a frag time is greather than a death time, reset the killstreak.